new file:   .vscode/launch.json 	renamed:    unit/web/cache.go -> app/cache.go 	new file:   app/cron.go 	new file:   app/var.go 	modified:   config/config.yaml 	modified:   go.mod 	modified:   go.sum 	modified:   log/run.log 	modified:   main.go 	modified:   main_test.go 	new file:   unit/check/location_check.go 	new file:   unit/check/var.go 	modified:   unit/config/config.go 	modified:   unit/getter/base.go 	modified:   unit/proxy/base.go 	modified:   unit/proxy/clash.go 	modified:   unit/proxy/var.go 	modified:   unit/proxy/vmess.go 	modified:   unit/tool/base.go 	new file:   unit/tool/http_client.go 	deleted:    unit/tool/httpclient.go 	renamed:    unit/web/router.go -> web/router.go

+++ unit/proxy/vmess.go
@@ -34,7 +34,7 @@ func (proxy *Proxy) vmessConstruct(group, ps, add string, port any, fakeType, id
 	} else {
 		proxy.UUID = id
 	}
-	proxy.AlterID = cast.ToUint16(aid)
+	proxy.AlterID = cast.ToInt(aid)
 	proxy.EncryptMethod = cipher
 	if net == "" {
 		proxy.TransferProtocol = "tcp"
